# RailWayStation/docs/swagger.yaml
basePath: /api
definitions:
  models.Area:
    properties:
      districts:
        items:
          $ref: '#/definitions/models.District'
        type: array
      id:
        type: integer
      name:
        type: string
      region:
        $ref: '#/definitions/models.Region'
      region_id:
        type: integer
    type: object
...
  models.District:
    properties:
      area:
        $ref: '#/definitions/models.Area'
      area_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      rail_stations:
        items:
          $ref: '#/definitions/models.RailStation'
        type: array
    type: object
...
  models.Line:
    properties:
      id:
        type: string
      name:
        type: string
      rail_stations:
        items:
          $ref: '#/definitions/models.RailStation'
        type: array
    type: object
  models.LineCreateRequest:
    properties:
      id:
        maxLength: 20
        minLength: 1
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - id
    - name
    type: object
  models.LineUpdateRequest:
    properties:
      name:
        maxLength: 100
        minLength: 2
        type: string
    type: object
  models.RailStation:
    properties:
      district:
        $ref: '#/definitions/models.District'
      district_id:
        type: integer
      hash:
        type: string
      id:
        type: integer
      lat:
        type: number
      line:
        $ref: '#/definitions/models.Line'
      line_id:
        type: string
      lon:
        type: number
      name:
        type: string
    type: object
...
  models.Region:
    properties:
      areas:
        items:
          $ref: '#/definitions/models.Area'
        type: array
      id:
        type: integer
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API для управления железнодорожными станциями
  title: Railway Stations API
  version: "1.0"
paths:
...
  /lines:
    get:
      description: Get a list of all railway lines with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Search by name or ID
        in: query
        name: search
        type: string
      - default: id
        description: Sort field
        in: query
        name: sort_by
        type: string
      - default: asc
        description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all lines
      tags:
      - lines
    post:
      consumes:
      - application/json
      description: Create a new railway line in the database
      parameters:
      - description: Line object
        in: body
        name: line
        required: true
        schema:
          $ref: '#/definitions/models.LineCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Line'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new line
      tags:
      - lines
  /lines/{id}:
    delete:
      description: Delete a railway line by ID
      parameters:
      - description: Line ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete line
      tags:
      - lines
    get:
      description: Get a single railway line by its ID
      parameters:
      - description: Line ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Line'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get line by ID
      tags:
      - lines
    put:
      consumes:
      - application/json
      description: Update an existing railway line by ID
      parameters:
      - description: Line ID
        in: path
        name: id
        required: true
        type: string
      - description: Line update object
        in: body
        name: line
        required: true
        schema:
          $ref: '#/definitions/models.LineUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Line'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update line
      tags:
      - lines
...
swagger: "2.0"